Option to set items per page, on a per model basis, v2
This builds on the previous PR for a global override of the number of items per page, and allows it to be provided on a per-model basis.

For example in app/upmin/models/admin_shipment.rb:

    class AdminShipment < Upmin::Model
      items_per_page 20
    end

If not specified, or no custom class exisits, then either the global override `config.items_per_page` or the builtin default will be used.